Add analytics impl

+++ frontend/frontend/src/app/components/admin/user-list/user-list.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { ToastrService } from 'ngx-toastr';
 import { User } from 'src/app/common/user';
 import { AdminUserService } from 'src/app/services/admin-user.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-user-list',
@@ -11,7 +13,7 @@ export class UserListComponent implements OnInit {
 
   users: User [] = [];
 
-  constructor(private adminUserService:AdminUserService){}
+  constructor(private adminUserService:AdminUserService, private toastr:ToastrService){}
 
   ngOnInit(): void {
     this.listUsers();
@@ -26,12 +28,30 @@ export class UserListComponent implements OnInit {
   }
 
   deleteUser(id: number) {
-    this.adminUserService.deleteUser(id).subscribe(
-      response => {
-        console.log("Usuario eliminado correctamente");
-        this.listUsers(); // Volver a cargar la lista
-      }
-    );
-  }
+    console.log('El ID de la categoria antes de eliminar es: '+id);
 
+    Swal.fire({
+      title: "¿Está seguro que desea eliminar el registro?",
+      text: "",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Eliminar",
+      cancelButtonText: "Cancelar",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.adminUserService.deleteUser(id).subscribe(
+          ()=> this.listUsers()
+        );
+        Swal.fire({
+          title: "¡Categorías!",
+          text: "Categoría eliminada correctamente.",
+          icon: "success"
+        });
+        this.listUsers();
+      }
+    });
+
+}
 }
